Standard Library Modules: Fix spurious `_Init_locks` dllexport (#3233)
