+++ stl/src/stacktrace.cpp
@@ -6,8 +6,6 @@
 // Do not include or define anything else here.
 // In particular, basic_string must not be included here.
 
-#include <yvals.h>
-
 #include <cstdio>
 #include <cstdlib>
 
@@ -166,7 +164,9 @@ namespace {
 
                 off = string_fill(fill, off + max_disp_num, str, [displacement, off](char* s, size_t) {
                     const int ret = std::snprintf(s + off, max_disp_num, "+0x%llX", displacement);
-                    _STL_VERIFY(ret > 0, "formatting error");
+                    if (ret <= 0) {
+                        std::abort(); // formatting error
+                    }
                     return off + ret;
                 });
             }
@@ -230,7 +230,9 @@ namespace {
 
                 off = string_fill(fill, off + max_line_num, str, [line, off](char* s, size_t) {
                     const int ret = std::snprintf(s + off, max_line_num, "(%u): ", line);
-                    _STL_VERIFY(ret > 0, "formatting error");
+                    if (ret <= 0) {
+                        std::abort(); // formatting error
+                    }
                     return off + ret;
                 });
             }
@@ -331,7 +333,9 @@ void __stdcall __std_stacktrace_to_string(const void* const* const _Addresses, c
 
         off = string_fill(_Fill, off + max_entry_num, _Str, [off, i](char* s, size_t) {
             const int ret = std::snprintf(s + off, max_entry_num, "%u> ", static_cast<unsigned int>(i));
-            _STL_VERIFY(ret > 0, "formatting error");
+            if (ret <= 0) {
+                std::abort(); // formatting error
+            }
             return off + ret;
         });
